RaspberryPi/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
entrypoints==0.3
gpiozero==1.5.1
grove.py==0.6
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4
python-dateutil==2.8.0
pyxdg==0.25
requests==2.21.0
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
smbus2==0.3.0
spidev==3.3
ssh-import-id==5.7
urllib3==1.24.1
